requirements.txt
amqp==5.1.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.4
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
Django==4.2.5
django-cors-headers==4.2.0
django-dirtyfields==1.9.2
django-filter==23.3
django-model-utils==4.3.1
django-soft-delete==0.9.21
djangorestframework==3.14.0
jwt==1.3.1
kombu==5.3.2
pipext==0.2.1
prompt-toolkit==3.0.39
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
ipython==8.12.3
django-environ==0.11.2
pyOpenSSL==23.2.0
certifi==2023.7.22
Pillow==10.1.0
